refactor(extension): Installer abstraction
